+++ atacseq/ccbr_trim_align/ChangeLog.md
@@ -27,3 +27,4 @@
   - bedtools **jaccard** with PCA is now in *ccbr_jaccard_pca.bash*
   - **pandoc** installed for rendering Rmd files
 
+#